ensure *taskMetadata type returned from GetTaskMetadataProvider has a stable interface going forward
